Refactor: [#77] role -> roles 복수 변경 및 common constant 생성 및 권한 상수 이동
