+++ src/common/constant.ts
@@ -1,0 +1,5 @@
+export const ROLES_KEY = {
+  ADMIN: 1,
+  MEMBER: 2,
+  USER: 3,
+};
