+++ src/common/decorators/roles.decorator.ts
@@ -1,0 +1,4 @@
+import { SetMetadata } from '@nestjs/common';
+import { ROLES_KEY } from '../constant';
+
+export const Roles = (role: number) => SetMetadata(ROLES_KEY, role);
